Update the installation scripts.

+++ scripts/uninstall-repo.sh
@@ -1,7 +1,11 @@
 #!/data/data/com.termux/files/usr/bin/env bash
 
 uninstall(){
-    rm -rf $PREFIX/etc/apt/sources.list.d/alyxs-tur.list
+    if [ -f "$PREFIX/etc/apt/sources.list.d/alyxs-tur.list"]; then
+        echo "You do not have this repository installed."
+    else
+        rm -rf $PREFIX/etc/apt/sources.list.d/alyxs-tur.list
+    fi
 }
 
 uninstall
